Refactor profiles/models.py into a module folder, in preparation for an additional OrganizationProfile class
