+++ pulseapi/profiles/models/__init__.py
@@ -1,0 +1,5 @@
+from .profiletype import ProfileType
+from .programtype import ProgramType
+from .programyear import ProgramYear
+from .userprofile import UserProfile, entry_thumbnail_path
+from .userbookmarks import UserBookmarks
